+++ swarm_copy/app/dependencies.py
@@ -13,12 +13,12 @@ from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncEngine, AsyncSession
 from starlette.status import HTTP_401_UNAUTHORIZED
 
+from swarm_copy.agent_routine import AgentsRoutine
 from swarm_copy.app.app_utils import validate_project
 from swarm_copy.app.config import Settings
 from swarm_copy.app.database.sql_schemas import Threads
 from swarm_copy.cell_types import CellTypesMeta
 from swarm_copy.new_types import Agent
-from swarm_copy.run import AgentsRoutine
 from swarm_copy.tools import (
     ElectrophysFeatureTool,
     GetMorphoTool,
@@ -204,9 +204,7 @@ async def get_vlab_and_project(
         if not thread:
             raise HTTPException(
                 status_code=404,
-                detail={
-                    "detail": "Thread not found.",
-                },
+                detail="Thread not found.",
             )
         if thread and thread.vlab_id and thread.project_id:
             vlab_and_project = {