+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CRUDs.
 - BlueNaas CRUD tools
 - Tests for threads module
+- app unit tests
+- Tests of AgentsRoutine.
 - Unit tests for database
 
 ### Fixed